requirements.txt
attrs==21.4.0
certifi==2022.6.15
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
cycler==0.11.0
Fiona==1.8.21
Flask==2.1.2
fonttools==4.33.3
geopandas==0.11.0
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
kiwisolver==1.4.3
MarkupSafe==2.1.1
matplotlib==3.5.2
munch==2.5.0
mysql-connector-python==8.0.29
numpy==1.23.0
packaging==21.3
pandas==1.4.3
Pillow==9.1.1
plotly==5.9.0
protobuf==4.21.2
pyparsing==3.0.9
pyproj==3.3.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
scikit-learn==1.1.1
scipy==1.8.1
Shapely==1.8.2
six==1.16.0
tenacity==8.0.1
threadpoolctl==3.1.0
Werkzeug==2.1.2
zipp==3.8.0
